requirements.txt
pycryptodome>=3.18.0
numpy
psutil
matplotlib
pytest
